@@ -1,23 +1,25 @@
 module github.com/cgascoig/isctl
 
-go 1.14
+go 1.18
 
 require (
-	github.com/CiscoDevNet/intersight-go v0.0.0-20230417035431-2e871659812f
 	github.com/PaesslerAG/jsonpath v0.1.1
 	github.com/bndr/gotabulate v1.1.3-0.20210209140143-7b841ac46153
-	github.com/fsnotify/fsnotify v1.4.9 // indirect
 	github.com/getkin/kin-openapi v0.85.0
-	github.com/kr/pretty v0.2.0 // indirect
-	github.com/mattn/go-runewidth v0.0.13 // indirect
 	github.com/mitchellh/go-homedir v1.1.0
-	github.com/sirupsen/logrus v1.2.0
-	github.com/spf13/afero v1.2.2 // indirect
+	github.com/sirupsen/logrus v1.8.1
 	github.com/spf13/cobra v1.4.0
 	github.com/spf13/pflag v1.0.5
-	github.com/spf13/viper v1.4.0
+	github.com/spf13/viper v1.8.1
 	github.com/stretchr/testify v1.7.0
-	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
 	gopkg.in/yaml.v2 v2.4.0
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
+)
+
+require (
+	github.com/cgascoig/intersight-simple-go v0.4.0 // indirect
+	github.com/fsnotify/fsnotify v1.4.9 // indirect
+	github.com/go-fed/httpsig v1.1.0 // indirect
+	github.com/icza/dyno v0.0.0-20230330125955-09f820a8d9c0 // indirect
+	github.com/mattn/go-runewidth v0.0.13 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )
