@@ -2,15 +2,13 @@ package main
 
 import (
 	"net/http"
-
-	openapi "github.com/CiscoDevNet/intersight-go"
 )
 
 // Operation is a generic interface for Intersight Operations
 type Operation interface {
-	SetBodyParams(*openapi.APIClient, map[string]interface{}) error
+	SetBodyParams(*isctlClient, map[string]interface{}) error
 
 	GetReferencedClasses(params map[string]interface{}) ([]string, error)
 
-	Execute(*openapi.APIClient, []string, map[string]string) (interface{}, *http.Response, error)
+	Execute(*isctlClient, []string, map[string]string) (interface{}, *http.Response, error)
 }
